LayoutTests/editing/find/find-and-replace-basic-expected.txt

Verifies that find and replace can be used to replace words in an editable area. This test requires CyberKitTestRunner.

After replacing 'orange' with 'apricot':
| "
        "
| <p>
|   "Apple banana <#selection-anchor>apricot<#selection-focus>."
| "
        "
| <p>
|   "Kiwi banana pear."
| "
    "

After replacing 'banana' with 'watermelon':
| "
        "
| <p>
|   "Apple <#selection-anchor>watermelon<#selection-focus> apricot."
| "
        "
| <p>
|   "Kiwi watermelon pear."
| "
    "
